pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "content-engine"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "openai-agents",
    "python-dotenv>=1.0.0",
]

[tool.setuptools.packages.find]
include = ["engine_agents*", "tools*", "schemas*", "flows*"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: live API tests, requires OPENAI_API_KEY (skipped by default)",
]
